ai-personal-assistance-journaling-app-23064115-python/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.22.1
certifi==2023.5.7
cffi==1.15.1
channels==3.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.6
constantly==15.1.0
cryptography==41.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==4.2.3
django-channels==0.7.0
django-redis==5.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
frozenlist==1.4.0
gTTS==2.3.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
qrcode==7.3.1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
yarl==1.9.2
zope.interface==6.0
